Adds the UrlBuilder class.
This is all about improving the handling of data from Atlas and the links
between objects. As we don't have any "id"'s for objects, the tag is the closest
unifying thing between items returned.

In theory, this should allow us to keep everything up in sync _and_ vastly
simplify the way we're building URL strings.