Implement validations for required attributes.
Boxes now have some attributes which are required and some which are
optional. This introduces an ActiveModel-like DSL, with `requires` that
allows you to specify those attributes which are required. When saving,
`validate!` can now be called on a resource which will check everything
required is there before making a request.